Add missing stuff for user theme image on error
Previous commit was missing "index.php" leading to a 404, and the fallback system image showing. Priority of the two images was also reversed.

Note: To be honest, I don't actually know what the ?v=1 / ?v=45 stuff does, but given the first one (?v=1 ) was already in the code, I assumed it was safe to use the second (?v=45) that generates the background image for the login page when using theming and reuse that path for the error pages.